my_site/pyproject.toml
[tool.poetry]
name = "test"
version = "1"
description = "1"
authors = ["1"]
license = "1"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "<5.0.0"
djangorestframework = "^3.14.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
